LWDEV-4027 Wip refactoring

+++ src/MarginTrading.MarketMaker/AzureRepositories/StorageModels/AssetPairExtPriceSettingsStorageModel.cs
@@ -7,7 +7,7 @@ namespace MarginTrading.MarketMaker.AzureRepositories.StorageModels
     internal class AssetPairExtPriceSettingsStorageModel
     {
         public string PresetDefaultExchange { get; set; }
-        public double OutlierThreshold { get; set; }
+        public decimal OutlierThreshold { get; set; }
         public TimeSpan MinOrderbooksSendingPeriod { get; set; }
         public MarkupsParams Markups { get; set; } = new MarkupsParams();
         public RepeatedOutliersParams RepeatedOutliers { get; set; } = new RepeatedOutliersParams();
@@ -17,15 +17,15 @@ namespace MarginTrading.MarketMaker.AzureRepositories.StorageModels
 
         public class MarkupsParams
         {
-            public double Bid { get; set; }
-            public double Ask { get; set; }
+            public decimal Bid { get; set; }
+            public decimal Ask { get; set; }
         }
 
         public class RepeatedOutliersParams
         {
             public int MaxSequenceLength { get; set; }
             public TimeSpan MaxSequenceAge { get; set; }
-            public double MaxAvg { get; set; }
+            public decimal MaxAvg { get; set; }
             public TimeSpan MaxAvgAge { get; set; }
         }
     }
