+++ src/MarginTrading.MarketMaker/Models/Api/AssetPairExtPriceSettingsModel.cs
@@ -11,7 +11,7 @@ namespace MarginTrading.MarketMaker.Models.Api
         public DateTimeOffset Timestamp { get; set; }
         public MarkupsModel Markups { get; set; }
         public string PresetDefaultExchange { get; set; }
-        public double OutlierThreshold { get; set; }
+        public decimal OutlierThreshold { get; set; }
         public RepeatedOutliersParamsModel RepeatedOutliers { get; set; }
         public IReadOnlyList<ExchangeExtPriceSettingsModel> Exchanges { get; set; }
         public ImmutableDictionary<OrderbookGeneratorStepEnum, bool> Steps { get; set; }
